requirements.txt
pdfplumber>=0.9.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
tqdm>=4.65.0
regex>=2023.8.8
python-dotenv>=1.0.0
google-genai>=0.2.0
pdf2image>=1.16.0
pytesseract>=0.3.10
